Filename Length Bugfix
This is a different attempt to fix the Windows file path limitation.

It won't help for any users running into issues with fringe custom paths,
but it should fix the issue for almost everybody.

This patch is yet untested in the wild, but I want to get some groundwork public.